update e2e testing - now passes

+++ tests/run-e2e-test.sh
@@ -2,39 +2,50 @@
 dir=$(dirname $0)
 cd ${dir}
 dir=$(pwd)
+echo "--> building operator container"
+export OPERATOR_TAG="kawaja.net/oaatoperator:e2e-$$"
+(
+   cd ..
+   docker build --file build/Dockerfile --tag ${OPERATOR_TAG} .
+)
+echo "--> updating python dependencies"
 python3 -m pip install --upgrade pip
 pip --quiet install -r ../requirements/dev.txt
-find e2e -name '*.yaml' -print0 | xargs -0 -n 1 yq . > /dev/null
-rm -f operator-output.log
 
-
-#(cd e2e
-#for dir in *; do
-#   if [ -d ${dir} ]; then
-#      ${dir}/update-steps.sh --cleanup
-#      ${dir}/update-steps.sh
-#   fi
-#done)
-#kubectl kuttl test
-#(cd e2e
-#for dir in *; do
-#   if [ -d ${dir} ]; then
-#      ${dir}/update-steps.sh --cleanup
-#   fi
-#done)
-
+echo "--> cleaning up old files"
 cd e2e
 for dir in *; do
    if [ -d ${dir} ]; then
-      k3d cluster create ${dir}x
-      docker pull busybox
-      k3d image import --cluster ${dir}x busybox
-      kubectl config use-context k3d-${dir}x
-      ${dir}/update-steps.sh --cleanup
-      ${dir}/update-steps.sh
-      sleep 60
-      (cd ..; kubectl kuttl test --timeout 240 --test ${dir})
-      k3d cluster delete ${dir}x
-      ${dir}/update-steps.sh --cleanup
+         ${dir}/update-steps.sh --cleanup
+   fi
+done
+
+echo "-->validating YAML files"
+find . -name '*.yaml' -print0 | xargs -0 -n 1 yq . > /dev/null
+rm -f operator-output.log
+
+echo "--> starting tests"
+for dir in *; do
+   if [ -d ${dir} ]; then
+      if [ -f "${dir}/update-steps.sh" ]; then
+         echo "**** setting up ${dir} test ****"
+         if [ ! -x "${dir}/update-steps.sh" ]; then
+            echo "file e2e/${dir}/update-steps.sh must be executable" >&2
+            exit 1
+         fi
+         k3d cluster delete ${dir}x || echo "delete failed (probably didn't exist), continuing"
+         k3d cluster create ${dir}x
+         docker pull busybox
+         k3d image import --verbose --cluster ${dir}x busybox ${OPERATOR_TAG}
+         kubectl config use-context k3d-${dir}x
+         kubectl cluster-info
+         ${dir}/update-steps.sh
+         echo "**** running ${dir} test ****"
+         sleep 60
+         (cd ..; kubectl kuttl test --timeout 240 --test ${dir})
+         echo "**** cleaning up ${dir} test ****"
+         k3d cluster delete ${dir}x
+         ${dir}/update-steps.sh --cleanup
+      fi
    fi
 done
